requirements.txt
alembic==1.0.10
cffi==1.12.3
Click==7.0
cryptography==1.7.2
enum34==1.1.6
Flask==1.0.3
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
passlib==1.7.1
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
redis==3.2.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.3
typing==3.5.3.0
Werkzeug==0.15.4
psycopg2==2.8.2
flask-restplus==0.12.1
elasticsearch==7.5.1
requests==2.22.0
